Fix theme not applying on body colors
